requirements.txt
dateparser
loguru
fire
rapidfuzz
newspaper3k
